@@ -1,38 +1,17 @@
 name: dslink
-version: 1.0.7
+version: 1.0.2
 description: "DSA IoT Platform - DSLink SDK for Dart"
 homepage: "http://iot-dsa.org"
 documentation: "https://iot-dsa.github.io/docs/sdks/dart/"
 dependencies:
-  args: "^0.13.4"
-  bignum: "^0.1.0"
-  dscipher: "^0.7.2"
-#  pointycastle: 1.0.2
-  logging: "^0.11.3"
-  crypto: ">=0.9.2 <3.0.0"
-  path: "^1.3.7"
-  msgpack: "^0.9.0"
-#  msgpack:
-#   git: https://github.com/tbelousova/msgpack.dart
-  json_diff: '^0.1.2'
-#  json_diff:
-#    git: https://github.com/tbelousova/dart-json_diff
-authors:
-  - Kenneth Endfinger <k.endfinger@dglogik.com>
-  - Rick Zhou <z.zhou@dglogik.com>
-  - Michael Bullington <m.bullington@dglogik.com>
-  - Matthew Butler <matthew.butler@acuitybrands.com>
-  - Joel Trottier-Hebert <joel.trottier-hebert@acuitybrands.com>
+  args: ^2.4.2
+  logging: ^1.2.0
+  crypto: ^3.0.3
+  path: ^1.9.0
+  msgpack_dart: ^1.0.1
+  pointycastle: ^3.7.3
+  json_diff: ^0.2.1
 dev_dependencies:
-  test: '0.12.15+8'
-#  test: any
-  dsbroker: "^1.1.1+1"
-#  browser: '^0.10.0'
-#  build_runner: any
-#  build_test: any
-#  build_web_compilers: any
+  test: ^1.5.3
 environment:
-#  sdk: '>=2.0.0'
-  sdk: '>=1.13.0 <2.0.0'
-executables:
-  dsbeacon: beacon
+  sdk: ">=3.0.0"
