@@ -1,9 +1,9 @@
 part of dslink.common;
 
 class TableColumn {
-  String type;
-  String name;
-  Object defaultValue;
+  String? type;
+  String? name;
+  Object? defaultValue;
 
   TableColumn(this.name, this.type, [this.defaultValue]);
 
@@ -33,7 +33,7 @@ class TableColumn {
   }
 
   /// parse List of Map into TableColumn
-  static List<TableColumn> parseColumns(List list) {
+  static List<TableColumn>? parseColumns(List list) {
     List<TableColumn> rslt = <TableColumn>[];
     for (Object m in list) {
       if (m is Map && m["name"] is String) {
@@ -56,7 +56,7 @@ class TableColumn {
 class Table {
   List<TableColumn> columns;
   List<List> rows;
-  Map meta;
+  Map? meta;
 
   Table(this.columns, this.rows, {this.meta});
 }
