@@ -1,16 +1,14 @@
-//FIXME:Dart1.0
-//*Dart1-open-block
 part of dslink.pk.dart;
 
-ECPrivateKey _cachedPrivate;
-ECPublicKey _cachedPublic;
+late ECPrivateKey _cachedPrivate;
+late ECPublicKey _cachedPublic;
 int _cachedTime = -1;
-String cachedPrivateStr;
+String? cachedPrivateStr;
 
 List<dynamic> generate(List<int> publicKeyRemote, String oldPriKeyStr) {
-  ECPoint publicPointRemote = _secp256r1.curve.decodePoint(publicKeyRemote);
-  ECPrivateKey privateKey;
-  ECPublicKey publicKey;
+  ECPoint? publicPointRemote = _secp256r1.curve.decodePoint(publicKeyRemote);
+  late ECPrivateKey privateKey;
+  late ECPublicKey publicKey;
   int ts = (new DateTime.now()).millisecondsSinceEpoch;
   if (cachedPrivateStr == null ||
       ts - _cachedTime > 60000 ||
@@ -19,14 +17,14 @@ List<dynamic> generate(List<int> publicKeyRemote, String oldPriKeyStr) {
     var gen = new ECKeyGenerator();
     var rsapars = new ECKeyGeneratorParameters(_secp256r1);
     var params = new ParametersWithRandom(rsapars,
-        DartCryptoProvider.INSTANCE.random);
+        DSRandomImpl());
     gen.init(params);
     var pair = gen.generateKeyPair();
-    privateKey = pair.privateKey;
-    publicKey = pair.publicKey;
+    privateKey = pair.privateKey as ECPrivateKey;
+    publicKey = pair.publicKey as ECPublicKey;
     if (oldPriKeyStr != '') {
-      _cachedPrivate = pair.privateKey;
-      _cachedPublic = pair.publicKey;
+      _cachedPrivate = pair.privateKey as ECPrivateKey;
+      _cachedPublic = pair.publicKey as ECPublicKey;
       _cachedTime = ts;
     }
   } else {
@@ -34,12 +32,11 @@ List<dynamic> generate(List<int> publicKeyRemote, String oldPriKeyStr) {
     publicKey = _cachedPublic;
   }
 
-  var Q2 = publicPointRemote * privateKey.d;
+  var Q2 = publicPointRemote! * privateKey.d;
   return [
-    /*privateKey.d.toByteArray()*/
-    bigIntegerToByteArray(privateKey.d),
-    publicKey.Q.getEncoded(false),
-    Q2.getEncoded(false)
+    bigIntToBytes(privateKey.d!),
+    publicKey.Q?.getEncoded(false),
+    Q2?.getEncoded(false)
   ];
 }
 
@@ -55,7 +52,7 @@ void _processECDH(SendPort initialReplyTo) {
 
 class ECDHIsolate {
   static bool get running => _ecdh_isolate != null;
-  static Isolate _ecdh_isolate;
+  static Isolate? _ecdh_isolate;
   static start() async {
     if (_ecdh_isolate != null) return;
     var response = new ReceivePort();
@@ -64,32 +61,32 @@ class ECDHIsolate {
     _checkRequest();
   }
 
-  static SendPort _isolatePort;
+  static late SendPort _isolatePort;
   static void _processResult(message) {
     if (message is SendPort) {
       _isolatePort = message;
     } else if (message is List) {
       if (_waitingReq != null && message.length == 3) {
-        var d1 = newBigIntegerFromBytes(1, message[0] as List<int>);
+        var d1 = readBytes(message[0]! as Uint8List);
         var Q1 = _secp256r1.curve.decodePoint(message[1] as List<int>);
         var Q2 = _secp256r1.curve.decodePoint(message[2] as List<int>);
         var ecdh = new ECDHImpl(
             new ECPrivateKey(d1, _secp256r1), new ECPublicKey(Q1, _secp256r1),
-            Q2);
-        _waitingReq._completer.complete(ecdh);
+            Q2!);
+        _waitingReq?._completer.complete(ecdh);
         _waitingReq = null;
       }
     }
     _checkRequest();
   }
 
-  static ECDHIsolateRequest _waitingReq;
+  static ECDHIsolateRequest? _waitingReq;
   static void _checkRequest() {
     if (_waitingReq == null && _requests.length > 0) {
       _waitingReq = _requests.removeFirst();
       _isolatePort.send([
-        _waitingReq.publicKeyRemote.ecPublicKey.Q.getEncoded(false),
-        _waitingReq.oldPrivate
+        _waitingReq!.publicKeyRemote.ecPublicKey.Q?.getEncoded(false),
+        _waitingReq?.oldPrivate
       ]);
     }
   }
@@ -99,8 +96,8 @@ class ECDHIsolate {
 
   /// when oldprivate is '', don't use cache
   static Future<ECDH> _sendRequest(
-      PublicKey publicKeyRemote, String oldprivate) {
-    var req = new ECDHIsolateRequest(publicKeyRemote, oldprivate);
+      PublicKey publicKeyRemote, String? oldprivate) {
+    var req = new ECDHIsolateRequest(publicKeyRemote as PublicKeyImpl, oldprivate);
     _requests.add(req);
     _checkRequest();
     return req.future;
@@ -109,11 +106,10 @@ class ECDHIsolate {
 
 class ECDHIsolateRequest {
   PublicKeyImpl publicKeyRemote;
-  String oldPrivate;
+  String? oldPrivate;
 
   ECDHIsolateRequest(this.publicKeyRemote, this.oldPrivate);
 
   Completer<ECDH> _completer = new Completer<ECDH>();
   Future<ECDH> get future => _completer.future;
 }
-//Dart1-close-block*/
