@@ -39,7 +39,7 @@ class RemoteNodeCache {
     _nodes.clear();
   }
 
-  Node getDefNode(String path, String defName) {
+  Node? getDefNode(String path, String defName) {
     if (DefaultDefNodes.nameMap.containsKey(defName)) {
       return DefaultDefNodes.nameMap[defName];
     }
@@ -47,17 +47,17 @@ class RemoteNodeCache {
   }
 
   /// update node with a map.
-  RemoteNode updateRemoteChildNode(RemoteNode parent, String name, Map m) {
+  RemoteNode? updateRemoteChildNode(RemoteNode parent, String name, Map m) {
     String path;
     if (parent.remotePath == '/') {
       path = '/$name';
     } else {
       path = '${parent.remotePath}/$name';
     }
-    RemoteNode rslt;
+    late RemoteNode? rslt;
     if (_nodes.containsKey(path)) {
       rslt = _nodes[path];
-      rslt.updateRemoteChildData(m, this);
+      rslt?.updateRemoteChildData(m, this);
     } else {
       rslt = new RemoteNode(path);
       _nodes[path] = rslt;
@@ -70,11 +70,11 @@ class RemoteNodeCache {
 class RemoteNode extends Node {
   final String remotePath;
   bool listed = false;
-  String name;
-  ListController _listController;
-  ReqSubscribeController _subscribeController;
+  late String name;
+  ListController? _listController;
+  ReqSubscribeController? _subscribeController;
 
-  ReqSubscribeController get subscribeController {
+  ReqSubscribeController? get subscribeController {
     return _subscribeController;
   }
 
@@ -83,12 +83,12 @@ class RemoteNode extends Node {
       return false;
     }
 
-    return _subscribeController._lastUpdate != null;
+    return _subscribeController?._lastUpdate != null;
   }
 
-  ValueUpdate get lastValueUpdate {
+  ValueUpdate? get lastValueUpdate {
     if (hasValueUpdate) {
-      return _subscribeController._lastUpdate;
+      return _subscribeController?._lastUpdate;
     } else {
       return null;
     }
@@ -122,14 +122,14 @@ class RemoteNode extends Node {
 
   /// whether the node's own data is updated
   bool isSelfUpdated() {
-    return _listController != null && _listController.initialized;
+    return _listController != null && _listController!.initialized;
   }
 
   Stream<RequesterListUpdate> _list(Requester requester) {
     if (_listController == null) {
       _listController = createListController(requester);
     }
-    return _listController.stream;
+    return _listController!.stream;
   }
 
   /// need a factory function for children class to override
@@ -141,17 +141,17 @@ class RemoteNode extends Node {
     if (_subscribeController == null) {
       _subscribeController = new ReqSubscribeController(this, requester);
     }
-    _subscribeController.listen(callback, qos);
+    _subscribeController?.listen(callback, qos);
   }
 
   void _unsubscribe(Requester requester, callback(ValueUpdate update)) {
     if (_subscribeController != null) {
-      _subscribeController.unlisten(callback);
+      _subscribeController?.unlisten(callback);
     }
   }
 
   Stream<RequesterInvokeUpdate> _invoke(Map params, Requester requester,
-    [int maxPermission = Permission.CONFIG, RequestConsumer fetchRawReq]) {
+    [int maxPermission = Permission.CONFIG, RequestConsumer? fetchRawReq]) {
     return new InvokeController(
       this,
       requester,
@@ -170,7 +170,7 @@ class RemoteNode extends Node {
       childPathPre = '$remotePath/';
     }
 
-    m.forEach((/*String*/ key, value) {
+    m.forEach((key, value) {
       if (key.startsWith(r'$')) {
         configs[key] = value;
       } else if (key.startsWith('@')) {
@@ -192,20 +192,20 @@ class RemoteNode extends Node {
     children.clear();
   }
 
-  Map save({bool includeValue: true}) {
+  Map save({bool includeValue = true}) {
     var map = {};
     map.addAll(configs);
     map.addAll(attributes);
     for (String key in children.keys) {
-      Node node = children[key];
-      map[key] = node is RemoteNode ? node.save() : node.getSimpleMap();
+      Node? node = children[key];
+      map[key] = node is RemoteNode ? node.save() : node?.getSimpleMap();
     }
 
     if (includeValue &&
       _subscribeController != null &&
-      _subscribeController._lastUpdate != null) {
-      map["?value"] = _subscribeController._lastUpdate.value;
-      map["?value_timestamp"] = _subscribeController._lastUpdate.ts;
+      _subscribeController?._lastUpdate != null) {
+      map["?value"] = _subscribeController?._lastUpdate?.value;
+      map["?value_timestamp"] = _subscribeController?._lastUpdate?.ts;
     }
 
     return map;
