@@ -1,12 +1,17 @@
 part of dslink.requester;
 
 /// request class handles raw response from responder
+/// Represents a request made by the requester.
+///
+/// This class encapsulates the necessary information for a request,
+/// including the requester instance, request ID, request data,
+/// and a callback for updating the request status.
 class Request {
   final Requester requester;
   final int rid;
-  final Map data;
+  final Map? data;
 
-  /// raw request callback
+  /// The callback function for updating the request status.
   final RequestUpdater updater;
   bool _isClosed = false;
   bool get isClosed => _isClosed;
@@ -15,22 +20,29 @@ class Request {
 
   String streamStatus = StreamStatus.initialize;
 
-  /// resend the data if previous sending failed
+  /// Resends the data if the previous sending failed.
+  ///
+  /// This method adds the data to the send list of the requester
+  /// to be resent later.
   void resend() {
-    requester.addToSendList(data);
+    requester.addToSendList(data!);
   }
-  
+
+  /// Adds request parameters to the send list.
+  ///
+  /// This method adds the request ID and parameters to the send list
+  /// of the requester to be sent later.
   void addReqParams(Map m) {
-    requester.addToSendList({'rid':rid, 'params':m});
+    requester.addToSendList({'rid': rid, 'params': m});
   }
 
   void _update(Map m) {
     if (m["stream"] is String) {
       streamStatus = m["stream"];
     }
-    List updates;
-    List columns;
-    Map meta;
+    List? updates;
+    List? columns;
+    Map? meta;
     if (m["updates"] is List) {
       updates = m["updates"];
     }
@@ -44,7 +56,7 @@ class Request {
     if (streamStatus == StreamStatus.closed) {
       requester._requests.remove(rid);
     }
-    DSError error;
+    DSError? error;
     if (m.containsKey("error") && m["error"] is Map) {
       error = new DSError.fromMap(m["error"]);
       requester._errorController.add(error);
@@ -53,17 +65,22 @@ class Request {
     updater.onUpdate(streamStatus, updates, columns, meta, error);
   }
 
-  /// close the request and finish data
-  void _close([DSError error]) {
+  /// Closes the request and finishes the data.
+  ///
+  /// This method sets the stream status to "closed" and calls the
+  /// updater callback with the updated status and optional error.
+  void _close([DSError? error]) {
     if (streamStatus != StreamStatus.closed) {
       streamStatus = StreamStatus.closed;
       updater.onUpdate(StreamStatus.closed, null, null, null, error);
     }
   }
 
-  /// close the request from the client side
+  /// Closes the request from the client side.
+  ///
+  /// This method is used to close the request from the client side.
+  /// It will also be called later from the requester.
   void close() {
-    // _close will also be called later from the requester;
     requester.closeRequest(this);
   }
 }
