@@ -1,14 +1,14 @@
 part of dslink.utils;
 
-typedef Object _Encoder(Object input);
-typedef Object _Reviver(String key, Object input);
+typedef Object? _Encoder(Object? input);
+typedef Object? _Reviver(Object? key, Object? input);
 
 class BinaryData {
   /// used when only partial data is received
   /// don"t merge them before it's finished
-  List<ByteData> mergingList;
+  List<ByteData>? mergingList;
 
-  ByteData bytes;
+  ByteData? bytes;
 
   BinaryData(ByteData bytes) {
     this.bytes = bytes;
@@ -27,23 +27,23 @@ abstract class DsCodec {
 
   static final DsCodec defaultCodec = DsJson.instance;
 
-  static void register(String name, DsCodec codec) {
+  static void register(String? name, DsCodec? codec) {
     if (name != null && codec != null) {
       _codecs[name] = codec;
     }
   }
 
   static DsCodec getCodec(String name) {
-    DsCodec rslt = _codecs[name];
+    DsCodec? rslt = _codecs[name];
     if (rslt == null) {
       return defaultCodec;
     }
     return rslt;
   }
 
-  Object _blankData;
+  Object? _blankData;
 
-  Object get blankData {
+  Object? get blankData {
     if (_blankData == null) {
       _blankData = encodeFrame({});
     }
@@ -51,18 +51,18 @@ abstract class DsCodec {
   }
 
   /// output String or List<int>
-  Object encodeFrame(Map val);
+  Object? encodeFrame(Map val);
 
   /// input can be String or List<int>
-  Map decodeStringFrame(String input);
+  Map? decodeStringFrame(String input);
 
-  Map decodeBinaryFrame(List<int> input);
+  Map? decodeBinaryFrame(List<int> input);
 }
 
 abstract class DsJson {
   static DsJsonCodecImpl instance = new DsJsonCodecImpl();
 
-  static String encode(Object val, {bool pretty: false}) {
+  static String encode(Object val, {bool pretty = false}) {
     return instance.encodeJson(val, pretty: pretty);
   }
 
@@ -70,7 +70,7 @@ abstract class DsJson {
     return instance.decodeJson(str);
   }
 
-  String encodeJson(Object val, {bool pretty: false});
+  String encodeJson(Object val, {bool pretty = false});
 
   dynamic decodeJson(String str);
 }
@@ -83,14 +83,14 @@ class DsJsonCodecImpl extends DsCodec implements DsJson {
   JsonEncoder encoder = new JsonEncoder(_safeEncoder);
 
   JsonDecoder decoder = new JsonDecoder();
-  JsonEncoder _prettyEncoder;
+  JsonEncoder? _prettyEncoder;
 
   dynamic decodeJson(String str) {
     return decoder.convert(str);
   }
 
-  String encodeJson(val, {bool pretty: false}) {
-    JsonEncoder e = encoder;
+  String encodeJson(val, {bool pretty = false}) {
+    JsonEncoder? e = encoder;
     if (pretty) {
       if (_prettyEncoder == null) {
         _prettyEncoder =
@@ -98,22 +98,22 @@ class DsJsonCodecImpl extends DsCodec implements DsJson {
       }
       e = _prettyEncoder;
     }
-    return e.convert(val);
+    return e!.convert(val);
   }
 
-  JsonDecoder _unsafeDecoder;
+  JsonDecoder? _unsafeDecoder;
 
-  Map decodeBinaryFrame(List<int> bytes) {
+  Map? decodeBinaryFrame(List<int> bytes) {
     return decodeStringFrame(const Utf8Decoder().convert(bytes));
   }
 
-  Map decodeStringFrame(String str) {
+  Map? decodeStringFrame(String str) {
     if (_reviver == null) {
       _reviver = (key, value) {
         if (value is String && value.startsWith("\u001Bbytes:")) {
           try {
             return ByteDataUtil.fromUint8List(
-                Base64.decode(value.substring(7)));
+                Base64.decode(value.substring(7))!);
           } catch (err) {
             return null;
           }
@@ -126,36 +126,35 @@ class DsJsonCodecImpl extends DsCodec implements DsJson {
       _unsafeDecoder = new JsonDecoder(_reviver);
     }
 
-    var result = _unsafeDecoder.convert(str);
+    var result = _unsafeDecoder?.convert(str);
     return result;
   }
 
-  _Reviver _reviver;
-  _Encoder _encoder;
+  _Reviver? _reviver;
+  _Encoder? _encoder;
 
-  Object encodeFrame(Object val) {
+  Object? encodeFrame(Object val) {
     if (_encoder == null) {
       _encoder = (value) {
         if (value is ByteData) {
-          return "\u001Bbytes:${Base64.encode(
-              ByteDataUtil.toUint8List(value))}";
+          return "\u001Bbytes:${Base64.encode(ByteDataUtil.toUint8List(value))}";
         }
         return null;
       };
     }
 
-    JsonEncoder c;
+    JsonEncoder? c;
 
     if (_unsafeEncoder == null) {
       _unsafeEncoder = new JsonEncoder(_encoder);
     }
     c = _unsafeEncoder;
 
-    var result = c.convert(val);
+    var result = c?.convert(val);
     return result;
   }
 
-  JsonEncoder _unsafeEncoder;
+  JsonEncoder? _unsafeEncoder;
 }
 
 class DsMsgPackCodecImpl extends DsCodec {
@@ -163,28 +162,24 @@ class DsMsgPackCodecImpl extends DsCodec {
 
   Map decodeBinaryFrame(List<int> input) {
     Uint8List data = ByteDataUtil.list2Uint8List(input);
-    if (_unpacker == null) {
-      _unpacker = new Unpacker(data.buffer, data.offsetInBytes);
-    } else {
-      _unpacker.reset(data.buffer, 0);
-      _unpacker.offset = data.offsetInBytes;
-    }
-    Object rslt = _unpacker.unpack();
+
+    _unpacker = new Deserializer(data);
+    
+    Object rslt = _unpacker?.decode();
     if (rslt is Map) {
       return rslt;
     }
-    _unpacker.data = null;
     return {};
   }
 
-  Unpacker _unpacker;
+  Deserializer? _unpacker;
 
   Map decodeStringFrame(String input) {
     // not supported
     return {};
   }
 
-  Object encodeFrame(Map val) {
-    return pack(val);
+  Object? encodeFrame(Map val) {
+    return serialize(val);
   }
 }
