@@ -2,9 +2,9 @@ import "dart:async";
 
 import "package:dslink/dslink.dart";
 
-LinkProvider link;
-int lastNum;
-SimpleNode valueNode;
+late LinkProvider link;
+late int lastNum;
+late SimpleNode valueNode;
 
 main(List<String> args) {
   var defaultNodes = <String, dynamic>{
@@ -22,7 +22,7 @@ main(List<String> args) {
   }
 
   link.connect();
-  link.onRequesterReady.then((Requester req) {
+  link.onRequesterReady.then((Requester? req) {
     Request rawreq;
     void fetchReq(Request v) {
       rawreq = v;
@@ -31,7 +31,7 @@ main(List<String> args) {
         rawreq.addReqParams({'Path':'/data/m1', 'Value':++i});
       });
     }
-    req.invoke(
+    req?.invoke(
       '/data/streamingSet', {'Path':'/data/m1', 'Value':0}, Permission.CONFIG,
       fetchReq).listen((update) {
       print(update.updates);
