@@ -11,9 +11,6 @@ import 'package:dslink/utils.dart';
 import "package:path/path.dart" as pathlib;
 import "package:crypto/crypto.dart";
 
-//FIXME:Dart2.0
-//import "convert_consts.dart";
-
 const bool _tcpNoDelay = const bool.fromEnvironment(
   "dsa.io.tcpNoDelay",
   defaultValue: true
@@ -43,7 +40,7 @@ class HttpHelper {
   static Future<HttpClientRequest> createRequest(
     String method,
     String url, {
-      Map<String, String> headers
+      Map<String, String>? headers
     }) async {
     var request = await client.openUrl(method, Uri.parse(url));
     if (headers != null) {
@@ -62,15 +59,15 @@ class HttpHelper {
 
   /// Fetches the specified [url] from HTTP.
   /// If [headers] is specified, the headers will be added onto the request.
-  static Future<String> fetchUrl(String url, {Map<String, String> headers}) async {
+  static Future<String> fetchUrl(String url, {Map<String, String>? headers}) async {
     var request = await createRequest("GET", url, headers: headers);
     var response = await request.close();
     return const Utf8Decoder().convert(await readBytesFromResponse(response));
   }
 
-  /// Fetches the specified [url] from HTTP as JSON.
+  /// Fetches the specified [url] from HTTP as json.
   /// If [headers] is specified, the headers will be added onto the request.
-  static Future<dynamic> fetchJSON(String url, {Map<String, String> headers}) async {
+  static Future<dynamic> fetchJSON(String url, {Map<String, String>? headers}) async {
     return const JsonDecoder().convert(await fetchUrl(url, headers: headers));
   }
 
@@ -83,12 +80,12 @@ class HttpHelper {
   /// Custom WebSocket Connection logic.
   static Future<WebSocket> connectToWebSocket(
       String url, {
-      Iterable<String> protocols,
-      Map<String, dynamic> headers,
-      HttpClient httpClient,
-      bool useStandardWebSocket
+      Iterable<String>? protocols,
+      Map<String, dynamic>? headers,
+      HttpClient? httpClient,
+      bool? useStandardWebSocket
     }) async {
-    Uri uri = Uri.parse(url);
+    Uri? uri = Uri.parse(url);
 
     if (useStandardWebSocket == null) {
       useStandardWebSocket = enableStandardWebSocket;
@@ -114,7 +111,7 @@ class HttpHelper {
     for (int i = 0; i < 16; i++) {
       nonceData[i] = random.nextInt(256);
     }
-    String nonce = BASE64.encode(nonceData);
+    String nonce = base64.encode(nonceData);
 
     int port = uri.port;
     if (port == 0) {
@@ -136,19 +133,19 @@ class HttpHelper {
     ) : httpClient;
 
     return _client.openUrl("GET", uri).then((HttpClientRequest request) async {
-      if (uri.userInfo != null && !uri.userInfo.isEmpty) {
+      if (uri?.userInfo != null && !uri!.userInfo.isEmpty) {
         // If the URL contains user information use that for basic
         // authorization.
-        String auth = BASE64.encode(UTF8.encode(uri.userInfo));
-        request.headers.set(HttpHeaders.AUTHORIZATION, "Basic $auth");
+        String auth = base64.encode(utf8.encode(uri.userInfo));
+        request.headers.set(HttpHeaders.authorizationHeader, "Basic $auth");
       }
       if (headers != null) {
         headers.forEach((field, value) => request.headers.add(field, value));
       }
       // Setup the initial handshake.
       request.headers
-        ..set(HttpHeaders.CONNECTION, "Upgrade")
-        ..set(HttpHeaders.UPGRADE, "websocket")
+        ..set(HttpHeaders.connectionHeader, "Upgrade")
+        ..set(HttpHeaders.upgradeHeader, "websocket")
         ..set("Sec-WebSocket-Key", nonce)
         ..set("Cache-Control", "no-cache")
         ..set("Sec-WebSocket-Version", "13");
@@ -166,19 +163,19 @@ class HttpHelper {
         });
         throw new WebSocketException(message);
       }
-      if (response.statusCode != HttpStatus.SWITCHING_PROTOCOLS ||
-        response.headers[HttpHeaders.CONNECTION] == null ||
-        !response.headers[HttpHeaders.CONNECTION].any(
+      if (response.statusCode != HttpStatus.switchingProtocols ||
+        response.headers[HttpHeaders.connectionHeader] == null ||
+        !response.headers[HttpHeaders.connectionHeader]!.any(
               (value) => value.toLowerCase() == "upgrade") ||
-        response.headers.value(HttpHeaders.UPGRADE).toLowerCase() != "websocket") {
+        response.headers.value(HttpHeaders.upgradeHeader)!.toLowerCase() != "websocket") {
         error("Connection to '$uri' was not upgraded to websocket");
       }
-      String accept = response.headers.value("Sec-WebSocket-Accept");
+      String? accept = response.headers.value("Sec-WebSocket-Accept");
       if (accept == null) {
         error("Response did not contain a 'Sec-WebSocket-Accept' header");
       }
       List<int> expectedAccept = sha1.convert("$nonce$_webSocketGUID".codeUnits).bytes;
-      List<int> receivedAccept = BASE64.decode(accept);
+      List<int> receivedAccept = base64.decode(accept!);
       if (expectedAccept.length != receivedAccept.length) {
         error("Response header 'Sec-WebSocket-Accept' is the wrong length");
       }
@@ -189,7 +186,7 @@ class HttpHelper {
       }
       var protocol = response.headers.value('Sec-WebSocket-Protocol');
       return response.detachSocket().then((socket) {
-        socket.setOption(SocketOption.TCP_NODELAY, _tcpNoDelay);
+        socket.setOption(SocketOption.tcpNoDelay, _tcpNoDelay);
         return new WebSocket.fromUpgradedSocket(
           socket,
           protocol: protocol,
@@ -202,86 +199,86 @@ class HttpHelper {
     });
   }
 
-  static Future<WebSocket> upgradeToWebSocket(HttpRequest request, [
-    protocolSelector(List<String> protocols),
-    bool useStandardWebSocket
-  ]) {
-    if (useStandardWebSocket == null) {
-      useStandardWebSocket = enableStandardWebSocket;
-    }
+static Future<WebSocket> upgradeToWebSocket(HttpRequest request, [
+  protocolSelector(List<String> protocols)?,
+  bool? useStandardWebSocket
+]) {
+  if (useStandardWebSocket == null) {
+    useStandardWebSocket = enableStandardWebSocket;
+  }
 
-    if (useStandardWebSocket) {
-      return WebSocketTransformer.upgrade(
-        request,
-        protocolSelector: protocolSelector
-      );
-    }
+  if (useStandardWebSocket) {
+    return WebSocketTransformer.upgrade(
+      request,
+      protocolSelector: protocolSelector
+    );
+  }
 
-    var response = request.response;
-    if (!WebSocketTransformer.isUpgradeRequest(request)) {
-      // Send error response.
-      response
-        ..statusCode = HttpStatus.BAD_REQUEST
-        ..close();
-      return new Future.error(
-        new WebSocketException("Invalid WebSocket upgrade request"));
-    }
+  var response = request.response;
+  if (!WebSocketTransformer.isUpgradeRequest(request)) {
+    // Send error response.
+    response
+      ..statusCode = HttpStatus.badRequest
+      ..close();
+    return new Future.error(
+      new WebSocketException("Invalid WebSocket upgrade request"));
+  }
 
-    Future<WebSocket> upgrade(String protocol) {
-      // Send the upgrade response.
-      response
-        ..statusCode = HttpStatus.SWITCHING_PROTOCOLS
-        ..headers.add(HttpHeaders.CONNECTION, "Upgrade")
-        ..headers.add(HttpHeaders.UPGRADE, "websocket");
-      String key = request.headers.value("Sec-WebSocket-Key");
-      String accept = BASE64.encode(
-        sha1.convert("$key$_webSocketGUID".codeUnits).bytes
-      );
-      response.headers.add("Sec-WebSocket-Accept", accept);
-      if (protocol != null) {
-        response.headers.add("Sec-WebSocket-Protocol", protocol);
-      }
-      response.headers.contentLength = 0;
-      return response.detachSocket()
-        .then((socket) {
-        socket.setOption(SocketOption.TCP_NODELAY, _tcpNoDelay);
+  Future<WebSocket> upgrade(String? protocol) {
+    // Send the upgrade response.
+    response
+      ..statusCode = HttpStatus.switchingProtocols
+      ..headers.add(HttpHeaders.connectionHeader, "Upgrade")
+      ..headers.add(HttpHeaders.upgradeHeader, "websocket");
+    String key = request.headers.value("Sec-WebSocket-Key")!;
+    String accept = base64.encode(
+      sha1.convert("$key$_webSocketGUID".codeUnits).bytes
+    );
+    response.headers.add("Sec-WebSocket-Accept", accept);
+    if (protocol != null) {
+      response.headers.add("Sec-WebSocket-Protocol", protocol);
+    }
+    response.headers.contentLength = 0;
+    return response.detachSocket()
+      .then((socket) {
+        socket.setOption(SocketOption.tcpNoDelay, _tcpNoDelay);
         return new WebSocket.fromUpgradedSocket(
           socket, protocol: protocol, serverSide: true);
       });
-    }
-
-    var protocols = request.headers['Sec-WebSocket-Protocol'];
-    if (protocols != null && protocolSelector != null) {
-      // The suggested protocols can be spread over multiple lines, each
-      // consisting of multiple protocols. To unify all of them, first join
-      // the lists with ', ' and then tokenize.
-      protocols = HttpHelper.tokenizeFieldValue(protocols.join(', '));
-      return new Future(() => protocolSelector(protocols)).then((/*String*/ protocol) {
-        if (protocols.indexOf(protocol) < 0) {
-          throw new WebSocketException(
-            "Selected protocol is not in the list of available protocols");
-        }
-        return protocol;
-      }).catchError((error) {
-        response
-          ..statusCode = HttpStatus.INTERNAL_SERVER_ERROR
-          ..close();
-        throw error;
-      }).then((result) {
-        if (result is String) {
-          return upgrade(result);
-        }
-        return null;
-      }).then((WebSocket socket) {
-        return socket;
-      });
-    } else {
-      return upgrade(null);
-    }
   }
 
+  var protocols = request.headers['Sec-WebSocket-Protocol'];
+  if (protocols != null && protocolSelector != null) {
+    // The suggested protocols can be spread over multiple lines, each
+    // consisting of multiple protocols. To unify all of them, first join
+    // the lists with ', ' and then tokenize.
+    protocols = HttpHelper.tokenizeFieldValue(protocols.join(', '));
+    return new Future(() => protocolSelector(protocols!).then((protocol) {
+      if (protocols!.indexOf(protocol) < 0) {
+        throw new WebSocketException(
+          "Selected protocol is not in the list of available protocols");
+      }
+      return protocol;
+    }).catchError((error) {
+      response
+        ..statusCode = HttpStatus.internalServerError
+        ..close();
+      throw error;
+    }).then((result) {
+      if (result is String) {
+        return upgrade(result);
+      }
+      return null;
+    }).then((WebSocket socket) {
+      return socket;
+    }));
+  } else {
+    return upgrade(null);
+  }
+}
+
   static List<String> tokenizeFieldValue(String headerValue) {
-    List<String> tokens = new List<String>();
+    List<String> tokens = <String>[];
     int start = 0;
     int index = 0;
     while (index < headerValue.length) {
@@ -300,7 +297,7 @@ class HttpHelper {
 
 /// Generates a random socket port.
 Future<int> getRandomSocketPort() async {
-  var server = await ServerSocket.bind(InternetAddress.LOOPBACK_IP_V4.address, 0);
+  var server = await ServerSocket.bind(InternetAddress.loopbackIPv4.address, 0);
   var port = server.port;
   await server.close();
   return port;
@@ -310,7 +307,7 @@ final _separator = pathlib.separator;
 
 Future<File> _safeWriteBase(File targetFile, dynamic content,
     Future<File> writeFunction(File file, dynamic content),
-    {bool verifyJson : false}) async {
+    {bool verifyJson = false}) async {
   final tempDirectory = await Directory.current.createTemp();
   final targetFileName = pathlib.basename(targetFile.path);
 
@@ -321,7 +318,7 @@ Future<File> _safeWriteBase(File targetFile, dynamic content,
   if (verifyJson) {
     final readContent = await tempFile.readAsString();
     try {
-      JSON.decode(readContent);
+      json.decode(readContent);
     } on FormatException catch (e, s) {
       canOverwriteOriginalFile = false;
       logger.severe(
@@ -341,7 +338,7 @@ Future<File> _safeWriteBase(File targetFile, dynamic content,
 }
 
 Future<File> safeWriteAsString(File targetFile, String content,
-    {bool verifyJson : false}) async {
+    {bool verifyJson = false}) async {
   return _safeWriteBase(
       targetFile, content,
       (File f, content) => f.writeAsString(content, flush: true),
@@ -349,7 +346,7 @@ Future<File> safeWriteAsString(File targetFile, String content,
 }
 
 Future<File> safeWriteAsBytes(File targetFile, List<int> content,
-    {bool verifyJson : false}) async {
+    {bool verifyJson = false}) async {
   return _safeWriteBase(
       targetFile, content,
       (File f, content) => f.writeAsBytes(content, flush: true),
