@@ -4,27 +4,12 @@ library dslink.responder;
 import "dart:async";
 import "dart:collection";
 import "dart:typed_data";
+import 'dart:convert';
+import 'package:pointycastle/block/aes_fast.dart';
+import 'package:pointycastle/api.dart';
 
 import "common.dart";
 import "utils.dart";
-
-//FIXME:Dart1.0
-//*Dart1-open-block
-import 'package:dscipher/block/aes_fast.dart';
-import 'package:dscipher/params/key_parameter.dart';
-import 'dart:convert';
-// Dart1-close-block*/
-
-//FIXME:Dart2.0
-/*Dart2-open-block
-import "package:dslink/convert_consts.dart";
-import "package:pointycastle/export.dart";
-Dart2-close-block*/
-/*
-import "package:pointycastle/block/aes_fast.dart";
-import "package:pointycastle/api.dart";
-*/
-
 
 part "src/responder/responder.dart";
 part "src/responder/response.dart";
