@@ -37,9 +37,9 @@ class Permission {
     'never': NEVER
   };
 
-  static int parse(Object obj, [int defaultVal = NEVER]) {
+  static int parse(Object? obj, [int defaultVal = NEVER]) {
     if (obj is String && nameParser.containsKey(obj)) {
-      return nameParser[obj];
+      return nameParser[obj]!;
     }
     return defaultVal;
   }
@@ -57,13 +57,13 @@ class PermissionList {
     for (Object obj in data) {
       if (obj is Map) {
         if (obj['id'] is String) {
-          idMatchs[obj['id']] = Permission.nameParser[obj['permission']];
+          idMatchs[obj['id']] = Permission.nameParser[obj['permission']]!;
         } else if (obj['group'] is String) {
           if (obj['group'] == 'default') {
-            defaultPermission = Permission.nameParser[obj['permission']];
+            defaultPermission = Permission.nameParser[obj['permission']]!;
           } else {
             groupMatchs[obj['group']] =
-                Permission.nameParser[obj['permission']];
+                Permission.nameParser[obj['permission']]!;
           }
         }
       }
@@ -72,7 +72,7 @@ class PermissionList {
 
   bool _FORCE_CONFIG = true;
 
-  int getPermission(Responder responder) {
+  int? getPermission(Responder responder) {
     // TODO Permission temp workaround before user permission is implemented
     if (_FORCE_CONFIG) {
       return Permission.CONFIG;
@@ -84,7 +84,7 @@ class PermissionList {
     int rslt = Permission.NEVER;
     for (String group in responder.groups) {
       if (groupMatchs.containsKey(group)) {
-        int v = groupMatchs[group];
+        int v = groupMatchs[group]!;
         if (v < rslt) {
           // choose the lowest permission from all matched group
           rslt = v;
