@@ -5,7 +5,7 @@ abstract class NodeProviderImpl extends NodeProvider {
 }
 
 abstract class LocalNodeImpl extends LocalNode {
-  LocalNodeImpl(String path) : super(path);
+  LocalNodeImpl(String? path) : super(path);
 
   Map serialize(bool withChildren) {
     var rslt = {};
@@ -47,7 +47,7 @@ abstract class LocalNodeImpl extends LocalNode {
       childPathPre = '$path/';
     }
 
-    m.forEach((/*String*/ key, value) {
+    m.forEach((key, value) {
       if (key.startsWith(r'$')) {
         configs[key] = value;
       } else if (key.startsWith('@')) {
@@ -67,8 +67,8 @@ abstract class LocalNodeImpl extends LocalNode {
     listChangeController.add(name);
   }
 
-  Response setAttribute(String name, Object value, Responder responder,
-      Response response) {
+  Response? setAttribute(String name, Object value, Responder responder,
+      Response? response) {
     if (!attributes.containsKey(name) || attributes[name] != value) {
       attributes[name] = value;
       updateList(name);
@@ -77,11 +77,11 @@ abstract class LocalNodeImpl extends LocalNode {
         (provider as SerializableNodeProvider).persist();
       }
     }
-    return response..close();
+    return response?..close();
   }
 
-  Response removeAttribute(String name, Responder responder,
-      Response response) {
+  Response? removeAttribute(String name, Responder responder,
+      Response? response) {
     if (attributes.containsKey(name)) {
       attributes.remove(name);
       updateList(name);
@@ -90,28 +90,28 @@ abstract class LocalNodeImpl extends LocalNode {
         (provider as SerializableNodeProvider).persist();
       }
     }
-    return response..close();
+    return response?..close();
   }
 
-  Response setConfig(String name, Object value, Responder responder,
-      Response response) {
+  Response? setConfig(String name, Object value, Responder responder,
+      Response? response) {
     var config = Configs.getConfig(name, profile);
-    response.close(config.setConfig(value, this, responder));
+    response?.close((config as ConfigSetting).setConfig(value, this, responder));
     return response;
   }
 
-  Response removeConfig(String name, Responder responder, Response response) {
+  Response? removeConfig(String name, Responder responder, Response? response) {
     var config = Configs.getConfig(name, profile);
-    return response..close(config.removeConfig(this, responder));
+    return response?..close((config as ConfigSetting).removeConfig(this, responder));
   }
 
-  Response setValue(
+  Response? setValue(
     Object value,
-    Responder responder,
-    Response response,
+    Responder? responder,
+    Response? response,
     [int maxPermission = Permission.CONFIG]) {
     updateValue(value);
     // TODO: check value type
-    return response..close();
+    return response?..close();
   }
 }
